requirements.txt
potassium
torch
transformers
google-cloud-storage